[#3] Get Keycloak Javascript adapter from a CDN, using SRI
In order to easier the development of the SPA Javascript integration
with Keycloak, we hardcoded the URL for the Keycloak JS adapter,
pointing to localhost, This won't work outside local development
environments. So we need to retrieve the JS adapter from a known,
accesible place. JsDeliver CND hosts a copy of such a library (we
verified that the hosted copy is 100% identical to the copy you get
from a Keycloak server). An added benefit is that we can use Sub
Resource Integrity (SRI[1]) for additional trust in the copy we get
from the CDN, This is exactly what we do with AWS Cognito Javascript
SDK.

[1] https://developer.mozilla.org/en-US/docs/Web/Security/Subresource_Integrity

Fixes #3