[#2] Externs files reorganization
A few weeks ago we decided to split the externs definition files into
individual files (one per library/external service). In fact, we do so
in other internal projects.

Right now we have Cognito externs in the "global" externs file. And
that file is in the same directory as the regular front-end files, and
not in a sub directory (./foreign-libs/externs/) like we do in other
internal projects.

Closes #2