Handle project names with hyphens
If we create a project with a name separated by hyphens e.g. test-app
in the index.html it will appear the same as the name instead of using
underscore. E.g., `<script>test-app.client.init(); </script>`. As this
is not valid Javascript, we need to change hyphens to underscores in
the right places.