Make Cognito session module refresh the ID token
AWS Cognito User Pools emits IT tokens with a short lifetime (1 hour
by default). This is a good idea in general, as JWT token revocation
is a problem (see [1] for details).

But this means we need to refresh the ID token before it expires. We
use a "half-lifetime refreshment" strategy[2]. We estimate the
remaining lifetime of the ID token when we get it and fire an
scheduled event (at half the remaining toke lifetime) to try and
refresh it. And if we are unable to refresh the token, we log the user
out (instead of silently failing to consume backend services because
we no longer have a valid ID token).

[1] https://www.oauth.com/oauth2-servers/access-tokens/access-token-lifetime/
[2] This is a common strategy in other domains, like DHCP lease renewals.