Make the front page HTML title configurable
Instead of using a hardcoded one (that we sometimes forget to change),
make it a mustache variable. If we don't post-process the project files
once the project is created, the "synthetic" title will be more
preeminent and easier to spot (and change).

If we use the internal Magnet project bootstrapping software, the
mustache variable will be post-processed and changed to whatever is
configured in the settings of the project.