Take info about keycloak process from url query
Previously we were using a custom cookie KEYCLOAK_PROCESS for that.
But that cookie was far from perfect. It was prone to many bugs like:
- If the user took too long to fill up the login form, the cookie
  would expire and after successful login nothing would happen.
- In scenario where user would decide to cancel login process and
  go back to landing page, he would be automatically redirected back
  to the login form. That effect would remain until the cookie dies.

This commit also removes `oksol.client.cookie` namespace, as it was
only used for the aformentioned cookie.

It also shifts responsibility for handling _wrong_ syntax of query
params returned by keycloak callback. Keycloak on callback simply
appends `state`, `session_state` and `code` params as a string looking
like this: `&state=...&session_state=...&code=...`.
That sometimes causes bugs because `secretary` expects the first param
to be introduced by `?` and not by `&`.
So now we just modify that url when the app loads.

It also adds auto token refreshment, therefore avoiding token
expiration.