Use Hikaricp db pool for application connections
Until now we were using the pool only for Ragtime migration, but not
for the application connections themselves. We thought it was done
automagically for us, but we were wrong. We were creating a new
connection each time we executed a query (and destroying it after the
query was done).

We need to explicitly use the Hikaricp connection object returned by
the Duct Hickaricp library. But we want to disable logging all the
queries executed by Hikaricp (as it fills the logs with thousands of
successful queries we don't care about). Thus we can't use
`:duct.module/sql` module, as it configures Hikaricp to use the
standar logger. We need to remove that module and configure Hikaricp
explicitly via `:duct.database.sql/hikaricp` key.

And while we are at it, we configure a slightly bigger connection pool
to cope for bigger concurrency loads (the default uses 10 as both the
maximum and mininum pool size)

[Fix #31]