Reuse existing Keycloak config var name for session config
Keycloak introduced some time ago a configuration variable named
KEYCLOAK_FRONTEND_URL to replace the deprecated variable (and removed
in recent verions) called KEYCLOAK_HOST. This new variable should
contain the official URL to access Keycloack from the external
systems ("front end" systems in Keycloak terminology).

That is exactly what we used our own KEYCLOAK_URL variable for. So
instead of creating another variable for this (and making sure we keep
them in sync) we simply reuse the standard Keycloak variable for this.