Fix Keycloak token lifetime and refresh interval calculation
According to https://openid.net/specs/openid-connect-core-1_0.html#IDToken,
regarding the `exp` claim, it states that:

    REQUIRED. Expiration time on or after which the ID Token MUST NOT
    be accepted for processing. The processing of this parameter
    requires that the current date/time MUST be before the expiration
    date/time listed in the value. [...] Its value is a JSON number
    representing the number of seconds from 1970-01-01T0:0:0Z as
    measured in UTC until the date/time.

This means we must compare the `exp` claim of the token with the
current time using the same metric. That is, the current time in
seconds from 1970-01-01T0:0:0Z as measured in UTC (aka Unix Epoch
time).

We are using `Date.prototype.getTime()`
(https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getTime)
to get the current time, which according to its documentation:

    The getTime() method returns the number of milliseconds since the
    Unix Epoch.

So we are using milliseconds, instead of seconds. This results in a
completely wrong value for the token remaining lifetime, and by
extension, the refresh interval value. We need to use the same metric
for both values.

[Re: #40]