Add missing changes for hydrogen/module.core upgrade to 4.0.2
9d83ef5 upgraded hydrogen/module.core to version 4.0.2. That version
allows the developers to specify a subset of figwheel-main
configuration options, using a map. Thus the default value in the
hydrogen/core profile was set to an empty map (to use the default
configuration options), instead of a boolean.

But we forgot to apply the same changes to the
hydrogen/session.cognito and hydrogen/session.keycloak profiles. The
reason we duplicate that configuration in these profiles is that Duct
doesn't allow us to perform merge or update operations on a given
config.edn Integrant key. Either we fully specify its content, or we
don't. Given that the session profiles are optional, we always need
to specify the value in hydrogen/core. But if we also enable the
session profile, we overwrite the value set by hydrogen/core, with
whatever value the session profile specifies. Thus we also need to
include the figwheel-main configuration option map in these cases.