Fix warnings from newer Google Closure compiler
It seems previous versions of the Google Closure compiler didn't care
about the case used in certain identifiers when declaring symbols as
externs, as long as the symbol names were right.

But newer versions are more strict in this respect and emit warnings
about the symbols not being defined before being used.

Simply using the same capitalization in all symbol names fixes the
issues.

[Re: #48]